# root/backend/swagger.yaml
openapi: 3.0.0
info:
  title: Nutrition API
  description: API pentru gestionarea aportului zilnic de calorii și produse nerecomandate.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
paths:
  /auth/register:
    post:
      summary: Înregistrare utilizator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
              required:
                - name
                - email
                - password
      responses:
        201:
          description: Utilizator înregistrat cu succes
        500:
          description: Eroare server

  /auth/login:
    post:
      summary: Autentificare utilizator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        200:
          description: Autentificare reușită
        404:
          description: Utilizator inexistent
        400:
          description: Credențiale invalide

  /auth/logout:
    post:
      summary: Deconectare utilizator
      responses:
        200:
          description: Deconectare reușită

  /products/search:
    get:
      summary: Căutare produs
      parameters:
        - name: query
          in: query
          description: Termenul de căutare pentru produs
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returnează produsele găsite

  /calories/public:
    post:
      summary: Obține aportul zilnic de kcal și lista de produse nerecomandate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                height:
                  type: number
                age:
                  type: number
                currentWeight:
                  type: number
                desiredWeight:
                  type: number
                bloodType:
                  type: integer
                  enum: [1, 2, 3, 4]
              required:
                - height
                - age
                - currentWeight
                - desiredWeight
                - bloodType
      responses:
        200:
          description: Returnează aportul zilnic recomandat de kcal și lista de produse nerecomandate

  /api/products:
    get:
      summary: Obține detaliile produsului
      description: Căutăm un produs după titlu pentru a obține informații despre acesta.
      operationId: getProductByTitle
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
          description: Titlul produsului pe care dorim să-l căutăm
      responses:
        '200':
          description: Produs găsit cu succes
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                  weight:
                    type: number
                  calories:
                    type: number
                  categories:
                    type: string
                  groupBloodNotAllowed:
                    type: array
                    items:
                      type: boolean
        '404':
          description: Produsul nu a fost găsit
        '500':
          description: Eroare la obținerea produsului

    post:
      summary: Adaugă un produs nou (doar local)
      description: Permite adăugarea unui produs în lista locală, fără a-l salva în baza de date.
      operationId: addProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                weight:
                  type: number
                calories:
                  type: number
                categories:
                  type: string
                groupBloodNotAllowed:
                  type: array
                  items:
                    type: boolean
      responses:
        '201':
          description: Produs adăugat cu succes
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                  weight:
                    type: number
                  calories:
                    type: number
                  categories:
                    type: string
                  groupBloodNotAllowed:
                    type: array
                    items:
                      type: boolean
        '400':
          description: Eroare la adăugarea produsului

    
  /products/{title}:
    delete:
      summary: Șterge un produs
      description: Șterge un produs pe baza titlului său din baza de date.
      operationId: deleteProduct
      parameters:
        - name: title
          in: path
          required: true
          schema:
            type: string
          description: Titlul produsului pe care dorim să-l ștergem
      responses:
        '200':
          description: Produsul a fost șters cu succes
        '400':
          description: Lipsă titlu produs
        '404':
          description: Produsul nu a fost găsit
        '500':
          description: Eroare la ștergerea produsului